implement integration test for Dilbert’s Quote Service & retrieving quote by id

+++ src/test/kotlin/net/loxal/soa/restkit/endpoint/ballot/PollResourceIT.kt
@@ -22,7 +22,7 @@ public class PollResourceIT : AbstractEndpointTest() {
     public fun createNewPoll() {
         val response = createPoll()
 
-        assertEquals(Response.Status.CREATED.getStatusCode().toLong(), response.getStatus().toLong())
+        assertEquals(Response.Status.CREATED.getStatusCode(), response.getStatus())
         assertEquals(MediaType.APPLICATION_JSON_TYPE, response.getMediaType())
         assertEquals(false, response.getLocation().getPath().endsWith("null"))
         assertEquals(true, response.getLocation().getPath().startsWith(Endpoint.URI_PATH_SEPARATOR + PollResource.RESOURCE_PATH))
@@ -38,7 +38,7 @@ public class PollResourceIT : AbstractEndpointTest() {
 
         assertEquals(MediaType.APPLICATION_JSON_TYPE, response.getMediaType())
         val notFoundStatus = Response.Status.NOT_FOUND
-        assertEquals(notFoundStatus.getStatusCode().toLong(), response.getStatus().toLong())
+        assertEquals(notFoundStatus.getStatusCode(), response.getStatus())
         val errorMessage = response.readEntity<ErrorMessage>(javaClass<ErrorMessage>())
         assertEquals(Response.Status.BAD_REQUEST.getReasonPhrase(), errorMessage.type)
     }
@@ -49,7 +49,7 @@ public class PollResourceIT : AbstractEndpointTest() {
 
         val deletion = AbstractEndpointTest.prepareTarget(existingPoll.getLocation()).request().delete()
 
-        assertEquals(Response.Status.NO_CONTENT.getStatusCode().toLong(), deletion.getStatus().toLong())
+        assertEquals(Response.Status.NO_CONTENT.getStatusCode(), deletion.getStatus())
         assertEquals(MediaType.APPLICATION_JSON_TYPE, deletion.getMediaType())
     }
 
@@ -59,7 +59,7 @@ public class PollResourceIT : AbstractEndpointTest() {
 
         val retrieval = AbstractEndpointTest.prepareTarget(existingPoll.getLocation()).request().get()
 
-        assertEquals(Response.Status.OK.getStatusCode().toLong(), retrieval.getStatus().toLong())
+        assertEquals(Response.Status.OK.getStatusCode(), retrieval.getStatus())
         assertEquals(MediaType.APPLICATION_JSON_TYPE, retrieval.getMediaType())
         val retrievedPoll = retrieval.readEntity<net.loxal.soa.restkit.model.ballot.Poll>(javaClass<Poll>())
         assertEquals(POLL_QUESTION, retrievedPoll.question)
@@ -76,7 +76,7 @@ public class PollResourceIT : AbstractEndpointTest() {
         val retrieval = AbstractEndpointTest.prepareTarget("${existingEntity.getLocation()} ${AbstractEndpointTest.NON_EXISTENT}").request().get()
 
 
-        assertEquals(Response.Status.NOT_FOUND.getStatusCode().toLong(), retrieval.getStatus().toLong())
+        assertEquals(Response.Status.NOT_FOUND.getStatusCode(), retrieval.getStatus())
         assertEquals(MediaType.APPLICATION_JSON_TYPE, retrieval.getMediaType())
         val notFoundError = retrieval.readEntity<ErrorMessage>(javaClass<ErrorMessage>())
         assertEquals(Response.Status.BAD_REQUEST.getReasonPhrase(), notFoundError.type)
@@ -96,13 +96,13 @@ public class PollResourceIT : AbstractEndpointTest() {
 
         val update = AbstractEndpointTest.prepareTarget(existingPoll.getLocation()).request().put(Entity.json<Poll>(modifiedPoll))
 
-        assertEquals(Response.Status.OK.getStatusCode().toLong(), update.getStatus().toLong())
+        assertEquals(Response.Status.OK.getStatusCode(), update.getStatus())
         assertEquals(MediaType.APPLICATION_JSON_TYPE, update.getMediaType())
 
         val retrievedUpdatedPoll = AbstractEndpointTest.prepareTarget(existingPoll.getLocation()).request().get()
         val updatedPoll = retrievedUpdatedPoll.readEntity<net.loxal.soa.restkit.model.ballot.Poll>(javaClass<Poll>())
         assertEquals(newQuestion, updatedPoll.question)
-        assertEquals(newAnswerOptions.size().toLong(), updatedPoll.answers.size().toLong())
+        assertEquals(newAnswerOptions.size().toInt(), updatedPoll.answers.size().toInt())
         assertEquals(firstAnswerOption, updatedPoll.answers.get(0))
         assertEquals(secondAnswerOption, updatedPoll.answers.get(1))
         assertEquals(thirdAnswerOption, updatedPoll.answers.get(2))
@@ -115,7 +115,7 @@ public class PollResourceIT : AbstractEndpointTest() {
         val somePoll = Poll("Irrelevant", listOf<String>())
         val update = AbstractEndpointTest.prepareTarget("${existingEntity.getLocation()} ${AbstractEndpointTest.NON_EXISTENT}").request().put(Entity.json<Poll>(somePoll))
 
-        assertEquals(Response.Status.NOT_FOUND.getStatusCode().toLong(), update.getStatus().toLong())
+        assertEquals(Response.Status.NOT_FOUND.getStatusCode(), update.getStatus())
         assertEquals(MediaType.APPLICATION_JSON_TYPE, update.getMediaType())
     }
 
@@ -127,7 +127,7 @@ public class PollResourceIT : AbstractEndpointTest() {
             val poll = Poll(POLL_QUESTION, POLL_ANSWERS)
 
             val createdPoll = AbstractEndpointTest.prepareGenericRequest(PollResource.RESOURCE_PATH).request().post(Entity.json<Poll>(poll))
-            assertEquals(Response.Status.CREATED.getStatusCode().toLong(), createdPoll.getStatus().toLong())
+            assertEquals(Response.Status.CREATED.getStatusCode(), createdPoll.getStatus())
             assertEquals(MediaType.APPLICATION_JSON_TYPE, createdPoll.getMediaType())
 
             return createdPoll
